A few simple speedups to the bundle builder (#156)
* Cache calculating CharmEndpointOptionality

Halves the amount of time spent calculating optionality

* Create ApplicationEndpoint only when needed

This single line change brings down the run-time of this function to 2/3
of original

* Refactor calculating non_optional_endpoints

Halved the time used in this function

* Use instance's cache instead of global one for is_optional

This one is actually faster than previous optimizations when the number
of unfulfilled endpoints, but roughly the same speed for larger number
of unfulfilled endpoints.

* Use stringform endpoint internally in the fn

Because creating the ApplicationEndpoint so many times ends up being
very expensive.  This change finally pushes the improvements to 2x the
original before this branch.

* Just use NamedTuple for ApplicationEndpoint, don't stringify

This is even faster!  But costs so much more memory :(

* Revert "Just use NamedTuple for ApplicationEndpoint, don't stringify"

This reverts commit 2061fa3c5f62fc9884e884a2ba89953d814021d2.

* Cleanup commented out code

* Add comment about lack of thread-safety

* Use tuple key instead of str key

* Add cached_method and use it for is_optional

* Improve type-hint annotation

as suggested by the copilot

* Fix formatting with double-quotes

* Remove unnecessary comment

* Replace walrus operator with simple impl